% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colonoscopy_setting.R
\name{colonoscopy_setting_codes}
\alias{colonoscopy_setting_codes}
\title{CPT/HCPCS codes defining the colonoscopy setting cohort}
\usage{
colonoscopy_setting_codes()
}
\value{
Character vector of CPT/HCPCS codes.
}
\description{
All CPT and HCPCS codes used to identify colonoscopy service lines in the
CMS physician/supplier data.
}
