% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget_impact.R
\name{estimate_budget_impact_all_comparators}
\alias{estimate_budget_impact_all_comparators}
\title{Budget impact of combined EMB vs. both alternatives}
\usage{
estimate_budget_impact_all_comparators(
  strategy_costs,
  cohort_sizes = c(10, 25, 50, 100, 1000)
)
}
\arguments{
\item{strategy_costs}{Tibble from
\code{compute_strategy_costs()$strategy_costs}.}

\item{cohort_sizes}{Numeric vector of annual patient volumes.}
}
\value{
A tibble binding \code{\link[=estimate_budget_impact]{estimate_budget_impact()}} results for
combined EMB vs. office EMB and vs. D&C.
}
\description{
Budget impact of combined EMB vs. both alternatives
}
