% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital_mrf_map.R
\name{plot_hospital_mrf_map}
\alias{plot_hospital_mrf_map}
\title{Static jittered point map of hospitals with usable MRF-derived payer-rate data}
\usage{
plot_hospital_mrf_map(hospital_points)
}
\arguments{
\item{hospital_points}{Tibble from \code{\link[=build_hospital_mrf_points]{build_hospital_mrf_points()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Static jittered point map of hospitals with usable MRF-derived payer-rate data
}
