% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_sensitivity.R
\name{summarize_geographic_sensitivity}
\alias{summarize_geographic_sensitivity}
\title{Summarize geographic sensitivity}
\usage{
summarize_geographic_sensitivity(geographic_analysis)
}
\arguments{
\item{geographic_analysis}{Return value from \code{\link[=run_geographic_sensitivity]{run_geographic_sensitivity()}}.}
}
\value{
List with \code{locality_summary} (tibble) and \code{summary_sentence} (character).
}
\description{
Summarize geographic sensitivity
}
