% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_sensitivity.R
\name{validate_geographic_inputs}
\alias{validate_geographic_inputs}
\title{Validate geographic sensitivity inputs}
\usage{
validate_geographic_inputs(
  locality_indices,
  pfs_rvus,
  professional_mapping,
  facility_mapping
)
}
\arguments{
\item{locality_indices}{Locality-level GPCI and wage-index table.}

\item{pfs_rvus}{CPT and place-of-service RVU table.}

\item{professional_mapping}{Mapping of model parameters to PFS RVUs.}

\item{facility_mapping}{Mapping of facility parameters to wage indices.}
}
\value{
Invisibly TRUE.
}
\description{
Validate geographic sensitivity inputs
}
